refactor(kernel): isolate caller resolution
